print numbers

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
 void print_number(int n)
 {
 unsigned int tens, digit, positive = n;
-double t_beg = 1;
+float t_beg = 1;
 
 if (n == 0)
 _putchar('0');
